move IntroductionSection component from js_modules
